Add custom path generator to zsh, add widget
Add a zsh function that automatically generates a short or long path
based on an environment variable. At the same time, add a widget that
makes use of this function, and bind the widget to a key combination.
Pressing this key combination would toggle the prompt displayed path
between a shortened path and the normal, long path.