Add tools update reminder script
Add the tools update reminder script, which is supposed to remind the
user on updates for the different tools installed. The tool is supposed
to be called by pacman/apt hooks on updates/installs, but of course can
be called on its own periodically.

For now, add fzf to the script as a tool.