Fix issue where neovim lua config files weren't recognized as lua
Since the neovim config files were not recognized as lua files (since, I
suppose, I only enable the filetype detection plugin in the said files),
add the modeline setting to set them as lua files.

The main issue with not being lua files was the 'textwidth', set as a
lua filetype option.