Change signcolumn option to 'auto'
This one's a weird one! First off, sidenote: a 'sign' takes 2
characters, so really 2 columns. So 2 signs will take 4 columns, 3 signs
6 columns and so on.

Change the 'signcolumn' option to automatically resize based on the
number of signs, with a minimum of 1 'sign' (so always display AT LEAST
2 columns for a sign, even if there is no sign), and a maximum of '4'
signs (so display AT MAXIMUM 8 columns worth of 4 signs).

The 'yes' option always keeps the sign column at 1 sign. The 'auto'
option toggles between 0 and 1 sign column, based on whether there are
signs to show. The 'yes:<num>' option has fixed columns, and the
'auto:<num>' resizes as needed up to num signs.

Seems like out of all, 'auto:<min>-<max>' really is the most versatile,
allowing the sign column to automatically jiggle between the min and max
numbers as needed.

This change basically increases the number of sign columns to be
displayed to be up to 4, whereas previously only 1 sign column was ever
displayed.