Move statusline.lua inside the lua/ dir
Since it's easire to reach the statusline setup, together with the other
lua-specific stuff.

Side-note: Initially, the idea of the lua/ directory was to contain only things
that are neovim-specific, with nothing related to plugins in here. The
idea of the after/plugin/ directory was to contain only things that are
plugin-specific, and all the keybinds, mappings, autocommands that
plugins need would be defined in there. The statusline is an exception
to both rules. It _mostly_ contains neovim-specific stuff, however the
nature of setting the status line up means that we will need extra stuff
from certain plugins (like version control diff stats, or treesitter
function scope). This makes it so that the statusline works with
plugin-specific stuff.

Since it cannot assume, by any means, what plugins are installed, it
works with a very thin interface, through Lua functions. Such functions
would need to be defined by plugins in after/plugin, and they are used
by the statusline in lua/statusline.lua.

For more info on this, see API in lua/statusline.lua.