Improve the statusline API
Make it clear what the API functions should return. This in turn
improves how the statusline treats the functions, and shouldn't resort
to hacky ways to call them.

In turn, this removes the need to check for valid treesitter parsers.
The function gets called nonetheless, and it returns what it considers
to be the correct result, based on the API of cursor_scope().