Add vim visual search plugin
Add a plugin to better treat cases where a visual selection is searched
via * and #. Neovim doesn't correctly treat cases where the pattern
contains slashes and other things which need to be escaped. The plugin
also adds <Leader>* and <Leader># to be able to search other files for
the same pattern.