Remove git branch information from statusline (temporary)
It seems that the git branch information is rather heavy, or something
else is happening with the vim.fn.system() call, as it constantly
flickers the cursor and other floating windows. Removing the system call
fixes all flickering, so disable this temporarily until I find a better
workaround to get branch information.