[wez] Fix ctrl-c and ctrl-v
Of course there had to be some bug I didn't think about related to
ctrl-c and ctrl-v in the terminal.

Make ctrl-c copy and clear selection. This allows a normal flow of "copy
this bit of text, but make ctrl-c free now, so you can normally ctrl-c
stuff in the terminal". Ctrl-shift-c and super-c still leave the
selection alive, though.

Ctrl-v interacts negatively with vim's visual mode block selection, so
that is removed for now in wezterm. The workaround is simply to not send
ctrl-v when the running program is neovim, and let neovim only bind
ctrl-v to paste in insert mode or term mode. See also neovim's help on
this (:h i_ctrl-v).